Break out utils file
Each function in its own file as a default export
