+++ src/components/SiteNavHeader.astro
@@ -1,5 +1,5 @@
 ---
-import { trimTrailingSlash } from "../utils/utils";
+import trimTrailingSlash from "../utils/trimTrailingSlash";
 import Logo from "./Logo.astro";
 
 const isCurrent = (path: string) =>
