+++ src/pages/feeds/feed.json.ts
@@ -1,6 +1,11 @@
 import metadata from '../../content/_metadata';
 import { getCollection } from 'astro:content';
-import { filterOutDraftsIfProduction, formatDateIso, getPostDate, sortByDate, renderMarkdown, getCategory } from '../../utils/utils';
+import formatDateIso from '../../utils/formatDateIso';
+import getPostDate from "../../utils/getPostDate";
+import sortByDate from "../../utils/sortByDate";
+import filterOutDraftsIfProduction from "../../utils/filterOutDraftsIfProduction";
+import getCategory from "../../utils/getCategory";
+import renderMarkdown from "../../utils/renderMarkdown";
 
 export async function GET(context) {
     const blog = sortByDate(filterOutDraftsIfProduction(await getCollection('posts')));
