+++ src/pages/og-images/_renderImage.ts
@@ -5,7 +5,7 @@ import sharp from "sharp";
 import styles from './_styles.css?raw';
 import faustinaRaw from '@fontsource/faustina/files/faustina-latin-400-normal.woff';
 import figtreeRaw from '@fontsource/figtree/files/figtree-latin-700-normal.woff';
-import { formatPostDate } from "../../utils/utils";
+import formatPostDate from "../../utils/formatPostDate";
 
 // They already are Buffers because of the custom rawFonts Vite plugin in astro.config.js, but TS doesn't know that
 const figtree = figtreeRaw as unknown as Buffer;
