+++ src/utils/countWords.ts
@@ -1,0 +1,6 @@
+// Adapted from Nunjuck's wordcount filter implementation
+
+export default function countWords(str: string): number {
+    const words = (str) ? str.match(/\w+/g) : null;
+    return (words) ? words.length : 0;
+}
