+++ src/utils/firstFive.ts
@@ -1,0 +1,4 @@
+
+export default function firstFive(collection: any[]) {
+    return collection.slice(0, 5);
+}
