+++ src/utils/formatDateIso.ts
@@ -1,0 +1,4 @@
+
+export default function formatDateIso(date: Date) {
+    return date.toISOString();
+}
