+++ src/utils/formatPostDate.ts
@@ -1,0 +1,5 @@
+import { DateTime } from "luxon";
+
+export default function formatPostDate(date: Date) {
+    return DateTime.fromJSDate(date).toLocaleString(DateTime.DATE_MED);
+}
