+++ src/utils/getPostDate.ts
@@ -1,0 +1,10 @@
+import { type CollectionEntry } from "astro:content";
+
+const now = new Date();
+
+export default function getPostDate(entry: CollectionEntry<'posts'>) {
+    if (entry.data.date) {
+        return entry.data.date;
+    }
+    else return now;
+}
