+++ src/utils/trimTrailingSlash.ts
@@ -1,0 +1,4 @@
+
+export default function trimTrailingSlash(str: string): string {
+    return str.replace(/\/$/, "");
+}
