+++ src/utils/unMarkdown.ts
@@ -1,0 +1,3 @@
+export default function unMarkdown(raw: string): string {
+    return raw?.replace(/[\*_`]/g, '')
+}
